requirements.txt
python-dotenv
ipykernel
langchain
langchain-openai
langchain-google-genai
langchain-community
streamlit